+++ packages/core/dlight/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @dlightjs/dlight
 
+## 1.0.0-beta.22
+
+### Patch Changes
+
+- refactor: change updateDerived to \_$ud
+
 ## 1.0.0-beta.21
 
 ### Patch Changes